- Batch-replaced "int *params" with "intprt_t *params" - Few fixes in interpreter.c's cast types - Changes in cmake to add or remove optional modules easier

+++ modules/libbgload/bgload.c
@@ -46,7 +46,7 @@
  * Helper function preparing params
  **/
 
-static bgdata *prep(int *params) {
+static bgdata *prep(intptr_t *params) {
     bgdata *t = (bgdata *)malloc(sizeof(bgdata));
     t->file   = strdup((char *)string_get(params[0]));
     t->id = (int *)params[1];
@@ -71,7 +71,7 @@ int bgDoLoad(void *d) {
 
 /* --------------------------------------------------------------------------- */
 
-int bgload(int (*fn)(), int *params) {
+int bgload(int (*fn)(), intptr_t *params) {
     bgdata *t = prep(params);
     t->fn     = fn;
